Moved file_index.py to parsers/.
