+++ python/fusion_engine_client/parsers/__init__.py
@@ -1,3 +1,4 @@
 from .encoder import FusionEngineEncoder
 from .decoder import FusionEngineDecoder
+from .file_index import FileIndex
 from .mixed_log_reader import MixedLogReader
